requirements.txt
anyio==3.6.2
certifi==2022.9.24
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
fastapi==0.89.1
filelock==3.9.0
h11==0.14.0
identify==2.5.17
idna==3.4
loguru==0.6.0
nodeenv==1.7.0
platformdirs==2.6.2
pre-commit==3.0.4
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.18.0
